Move generic functions to common package
