+++ tests/rpmlint/distgit-prepare.py
@@ -9,92 +9,18 @@
 import argparse
 import logging
 import os
-import re
-import sys
 import subprocess
-import tomllib
 from pathlib import Path
 from typing import Any
 
 import tomli_w
-from ruamel.yaml import YAML
+
+import utils
 
 logging.basicConfig(level="INFO")
 logger = logging.getLogger(Path(__file__).name)
 
 CI_CONFIG_SECTION = "rpmlint"
-CI_CONFIG_FILES = [
-    "ci.yaml",
-    "ci.yml",
-    "ci.toml",
-    "fedora-ci.yaml",
-    "fedora-ci.yml",
-    "fedora-ci.toml",
-]
-
-
-def get_config_from_ci_yaml(dist_git_path: Path) -> dict[str, Any] | None:
-    for ci_file_name in CI_CONFIG_FILES:
-        ci_file = dist_git_path / ci_file_name
-        if ci_file.exists():
-            break
-    else:
-        return None
-
-    logger.info(f"Found config file {ci_file_name}")
-    with ci_file.open("rb") as f:
-        if ci_file.suffix == ".toml":
-            full_config = tomllib.load(f)
-        else:
-            full_config = YAML().load(f)
-
-    if not (tools := full_config.get("tools")):
-        logger.info("No `tools` section found")
-        return None
-    if not (config := tools.get(CI_CONFIG_SECTION)):
-        logger.info(f"No `tools.{config}` section found")
-        return None
-    return config
-
-
-def get_dist_git(koji_task_id: str, workdir: Path) -> Path:
-    result = subprocess.run(
-        [
-            "koji",
-            "taskinfo",
-            "-v",
-            koji_task_id,
-        ],
-        capture_output=True,
-        text=True,
-        check=True,
-    )
-    task_info = result.stdout
-    task_error = result.stderr
-    logger.info(f"Task info output:\n{task_info}\nTask error:\n{task_error}")
-    source_match_obj = re.search(r"Source:\s*(.*)", task_info)
-    if source_match_obj is None:
-        logger.error(
-            "Could not find 'Source:' in koji taskinfo output. Maybe a 500 error? Please retry."
-        )
-        sys.exit(1)
-    source = source_match_obj.group(1)
-    source_match = re.search(r"git\+(?P<url>.*)#(?P<ref>.*)", source)
-    repo_url = source_match.group("url")
-    repo_ref = source_match.group("ref")
-
-    # Clone the dist-git used in the build
-    dist_git_path = workdir / "dist-git"
-    subprocess.run(
-        ["git", "clone", repo_url, dist_git_path],
-        check=True,
-    )
-    subprocess.run(
-        ["git", "checkout", "-d", repo_ref],
-        cwd=dist_git_path,
-        check=True,
-    )
-    return dist_git_path
 
 
 def set_config_files(config: dict[str, Any], args: argparse.Namespace) -> None:
@@ -132,9 +58,9 @@ def main(args: argparse.Namespace) -> None:
     """
     Prepare for rpmlint from a dist-git
     """
-    dist_git_path = get_dist_git(args.koji_task_id, args.workdir)
+    dist_git_path = utils.get_dist_git(args.koji_task_id, args.workdir)
 
-    if config := get_config_from_ci_yaml(dist_git_path):
+    if config := utils.get_config(dist_git_path, CI_CONFIG_SECTION):
         set_config_files(config, args)
     else:
         get_config_fallback(dist_git_path, args)
