+++ tests/rpmlint/utils
@@ -1,0 +1,1 @@
+../../utils
