+++ utils/__init__.py
@@ -1,0 +1,84 @@
+import logging
+import re
+import sys
+import subprocess
+import tomllib
+from pathlib import Path
+from typing import Any
+
+logger = logging.getLogger("tmt_plans.utils")
+
+CI_CONFIG_FILES = [
+    "ci.yaml",
+    "ci.yml",
+    "ci.toml",
+    "fedora-ci.yaml",
+    "fedora-ci.yml",
+    "fedora-ci.toml",
+]
+
+
+def get_config(dist_git_path: Path, section: str) -> dict[str, Any] | None:
+    from ruamel.yaml import YAML
+
+    for ci_file_name in CI_CONFIG_FILES:
+        ci_file = dist_git_path / ci_file_name
+        if ci_file.exists():
+            break
+    else:
+        return None
+
+    logger.info(f"Found config file {ci_file_name}")
+    with ci_file.open("rb") as f:
+        if ci_file.suffix == ".toml":
+            full_config = tomllib.load(f)
+        else:
+            full_config = YAML().load(f)
+
+    if not (tools := full_config.get("tools")):
+        logger.info("No `tools` section found")
+        return None
+    if not (config := tools.get(section)):
+        logger.info(f"No `tools.{config}` section found")
+        return None
+    return config
+
+
+def get_dist_git(koji_task_id: str, workdir: Path) -> Path:
+    result = subprocess.run(
+        [
+            "koji",
+            "taskinfo",
+            "-v",
+            koji_task_id,
+        ],
+        capture_output=True,
+        text=True,
+        check=True,
+    )
+    task_info = result.stdout
+    task_error = result.stderr
+    logger.info(f"Task info output:\n{task_info}\nTask error:\n{task_error}")
+    source_match_obj = re.search(r"Source:\s*(.*)", task_info)
+    if source_match_obj is None:
+        logger.error(
+            "Could not find 'Source:' in koji taskinfo output. Maybe a 500 error? Please retry."
+        )
+        sys.exit(1)
+    source = source_match_obj.group(1)
+    source_match = re.search(r"git\+(?P<url>.*)#(?P<ref>.*)", source)
+    repo_url = source_match.group("url")
+    repo_ref = source_match.group("ref")
+
+    # Clone the dist-git used in the build
+    dist_git_path = workdir / "dist-git"
+    subprocess.run(
+        ["git", "clone", repo_url, dist_git_path],
+        check=True,
+    )
+    subprocess.run(
+        ["git", "checkout", "-d", repo_ref],
+        cwd=dist_git_path,
+        check=True,
+    )
+    return dist_git_path
